Karlsbranch

+++ app/src/main/java/com/csis4495_cmk/webuy/models/Group.java
@@ -137,7 +137,7 @@ public class Group {
     }
 
 
-    public double getTax() {
+    public int getTax() {
         return tax;
     }
